fix: repair defects the new family fixtures and the options ratchet exposed
- Options surface: the eight text conditioners gained hiddenSize, numLayers,
  numHeads (and numKvHeads) constructor parameters, growing the ratchet gap
  from 977 to 1003. They now read a TextConditionerOptions, kept as a private
  copy so clones rebuild at the same size; null keeps each variant's value.
- MiniCPM-V/-o: scaled test options left DecoderDim 72 with 32 heads, so the
  decoder attention was built 64 wide. GenerativeVLMOptions now declares
  DecoderDim divisible by NumHeads, which the scaler realigns.
- DETR: the generator registered the throwing EnsureNeck/EnsureBackbone
  accessors as components, and a detector without a neck read as
  ShapeDeferred. They are [ParameterAlias] now, Neck is optional, and a
  declared-optional single component gets optional: true as adapters do.
- Safety: Evaluate(null) threw NullReferenceException where the audio and
  multimodal bases, ImageWatermarker, AudioWatermarker and
  TextImageAlignmentChecker skipped the guard the other bases have. The
  threshold-gated image fixtures now use a zero threshold and 32x32 content,
  and MultimodalGuardrail declares that a bare vector yields no findings.
- GP: MultiOutputGaussianProcess.Fit now fits the one-output case, so its
  integration test asserts that instead of the old throw.
- CodeQL: widen integer products before converting to double.

+++ src/MetaLearning/Algorithms/MetaDMAlgorithm.cs
@@ -374,7 +374,7 @@ public partial class MetaDMAlgorithm<T, TInput, TOutput> : MetaLearnerBase<T, TI
                 dot += syntheticFeatures[s][d] * NumOps.ToDouble(prototype[d]);
             agreement += dot;
         }
-        agreement /= syntheticFeatures.Length * _prototypeDim;
+        agreement /= (double)syntheticFeatures.Length * _prototypeDim;
 
         // Scale gradient: higher agreement → more confidence in gradient direction
         double scale = 1.0 + 0.1 * Math.Tanh(agreement);
